Moved mediaItem registration code to Schema/Types/Connection
