+++ src/Schema/Schema.php
@@ -9,6 +9,7 @@ class Schema {
 		new \WPGraphQLGutenberg\Schema\Types\InterfaceType\Block();
 		new \WPGraphQLGutenberg\Schema\Types\InterfaceType\BlockEditorContentNode();
 		new \WPGraphQLGutenberg\Schema\Types\Connection\BlockEditorContentNodeConnection();
+		new \WPGraphQLGutenberg\Schema\Types\Connection\CoreImageBlockMediaItemConnection();
 		new \WPGraphQLGutenberg\Schema\Types\BlockTypes();
 	}
 }
